# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job-Search Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get list of all locations",
                "description": "Returns list of all locations",
                "operationId": "Get all locations",
                "responses": {
                    "200": {
                        "description": "All Locations"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/locations/create": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create new location",
                "description": "Add new location",
                "operationId": "create new location",
                "requestBody": {
                    "description": "Enter location data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country",
                                    "city"
                                ],
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "format": "country",
                                        "example": "Egypt"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "city",
                                        "example": "Cairo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/locations/show": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "show specific location",
                "description": "show specific location",
                "operationId": "show specific location",
                "parameters": [
                    {
                        "name": "location_id",
                        "in": "query",
                        "description": "Enter location id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/locations/update": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update location",
                "description": "Edit location",
                "operationId": "update location",
                "requestBody": {
                    "description": "Enter location data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_id",
                                    "country",
                                    "city"
                                ],
                                "properties": {
                                    "location_id": {
                                        "type": "integer",
                                        "format": "location_id",
                                        "example": "1"
                                    },
                                    "country": {
                                        "type": "string",
                                        "format": "country",
                                        "example": "Egypt"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "city",
                                        "example": "Cairo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/locations/delete": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Soft delete location",
                "description": "Soft delete location",
                "operationId": "delete specific location",
                "requestBody": {
                    "description": "Pass location data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_id"
                                ],
                                "properties": {
                                    "location_id": {
                                        "type": "integer",
                                        "format": "location_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/locations/restore": {
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Restore delete location",
                "description": "Restore location",
                "operationId": "restore specific location",
                "requestBody": {
                    "description": "Pass location data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location_id"
                                ],
                                "properties": {
                                    "location_id": {
                                        "type": "integer",
                                        "format": "location_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "location restored successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}